increase maximum frame size to support newer cameras
Increase maximum frame size from 40,000 bytes to 200,000 bytes to support
newer high resolution cameras which can produce larger frames.

Without this change in place the parser fails to detect the start and end
frame markers if the frame size happens to be larger than 40,000 bytes.

Surely there is a better way of doing this, but this was the easiest way to
get around the problem.